factories

+++ core/src/main/java/entidades/Aliado.java
@@ -4,7 +4,7 @@ public class Aliado extends Personagem {
     private int nivel;
     private int experiencia;
 
-    public Aliado(String nome, int vidaMaxima, int defesa, int iniciativa, int nivel, int experiencia) {
+    public Aliado(String nome, int vidaMaxima, float defesa, int iniciativa, int nivel, int experiencia) {
         super(nome, vidaMaxima, defesa, iniciativa);
         this.nivel = nivel;
         this.experiencia = experiencia;
