+++ core/src/main/java/entidades/Inimigo.java
@@ -5,7 +5,7 @@ import progressao.Recompensa;
 public class Inimigo extends Personagem {
     private Recompensa recompensa;
 
-    public Inimigo(String nome, int vidaMaxima, int defesa, int iniciativa, Recompensa recompensa) {
+    public Inimigo(String nome, int vidaMaxima, float defesa, int iniciativa, Recompensa recompensa) {
         super(nome, vidaMaxima, defesa, iniciativa);
         this.recompensa = recompensa;
     }
